requirements.txt
python-dotenv
openai
gradio
json